requirements.txt
alpaca-py==0.40.0
annotated-types==0.7.0
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
exchange_calendars==4.10.1
executing==2.2.0
fonttools==4.58.0
frozendict==2.4.6
holidays==0.76
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
korean-lunar-calendar==0.3.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
msgpack==1.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pandas_market_calendars==5.1.1
parso==0.8.4
patsy==1.0.1
peewee==3.18.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.2
pyluach==2.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==27.0.0
requests==2.32.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
simfin==1.0.1
six==1.17.0
soupsieve==2.7
sseclient-py==1.8.0
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
yfinance==0.2.64
